ADdressing feedback from @Gryhon
You register your cog twice. first time in bot.py and secondly in the registeration.py class itself.
You have already a database connection inheritance by DiscordBase no need for a additional database connection in the class
No separation between business logic and data storage. The main reason to collect all database operations in DiscordBase is by design and to maintain database changes easily.
There will be more functions that needed a registration. therefore I will change the tablename and columns. That will be an issue, if you don't implement the database functions in DiscordBase.
I am very happy to see the update registration as an transaction. Good stuff! The chance that the blocking database by the transaction causes an issue should be very rare, but it could be an conflict with the reminder vote reminder. But as I said this conflict is nearly impossible.
If you are running the software native you will (could) become an issue using ctrl+c To ensure to handle that is the implementation of on_ready and background_task